# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AYA Assistant API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /children:
    get:
      operationId: listChildren
      tags: [children]
      summary: List children in the family
      responses:
        "200":
          description: List of children
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Child"
    post:
      operationId: createChild
      tags: [children]
      summary: Create a child profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChildBody"
      responses:
        "201":
          description: Child created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
  /children/{id}:
    get:
      operationId: getChild
      tags: [children]
      summary: Get a child by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Child
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateChild
      tags: [children]
      summary: Update a child profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChildBody"
      responses:
        "200":
          description: Updated child
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
    delete:
      operationId: deleteChild
      tags: [children]
      summary: Delete a child profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Chat ──────────────────────────────────────────────────────────────────
...
  /missions:
    get:
      operationId: listMissions
      tags: [missions]
      summary: List missions for a child
      parameters:
        - name: childId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Missions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mission"
  /missions/{id}/complete:
    post:
      operationId: completeMission
      tags: [missions]
      summary: Mark a mission as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mission completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Calendar ──────────────────────────────────────────────────────────────
...
  /family-tasks:
    get:
      operationId: listFamilyTasks
      tags: [tasks]
      summary: List family tasks
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyTask"
    post:
      operationId: createFamilyTask
      tags: [tasks]
      summary: Create a family task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFamilyTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyTask"
  /family-tasks/{id}:
    patch:
      operationId: updateFamilyTask
      tags: [tasks]
      summary: Update a family task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFamilyTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyTask"
    delete:
      operationId: deleteFamilyTask
      tags: [tasks]
      summary: Delete a family task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Progress ──────────────────────────────────────────────────────────────
  /progress:
    get:
      operationId: listProgress
      tags: [progress]
      summary: List learning progress for a child
      parameters:
        - name: childId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Progress"
    post:
      operationId: createProgress
      tags: [progress]
      summary: Record a progress entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgressBody"
      responses:
        "201":
          description: Progress recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    CreateChildBody:
      type: object
      properties:
        name:
          type: string
        grade:
          type: integer
          minimum: 1
          maximum: 12
        language:
          type: string
        country:
          type: string
        avatar:
          type: ["string", "null"]
      required: [name, grade, language, country]
    UpdateChildBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        grade:
          type: ["integer", "null"]
        language:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        xp:
          type: ["integer", "null"]
        stars:
          type: ["integer", "null"]
    Child:
      type: object
      properties:
        id:
          type: integer
        familyId:
          type: integer
        name:
          type: string
        grade:
          type: integer
        language:
          type: string
        country:
          type: string
        avatar:
          type: ["string", "null"]
        xp:
          type: integer
        stars:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, familyId, name, grade, language, country, avatar, xp, stars, createdAt]

    # ─── Chat ────────────────────────────────────────────────────────────────
...
    Mission:
      type: object
      properties:
        id:
          type: integer
        childId:
          type: integer
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        xpReward:
          type: integer
        starReward:
          type: integer
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, childId, title, description, subject, xpReward, starReward, completed, completedAt, createdAt]

    # ─── Calendar ────────────────────────────────────────────────────────────
...
    CreateFamilyTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        dueAt:
          type: ["string", "null"]
          format: date-time
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        assignedTo:
          type: ["integer", "null"]
      required: [title]
    UpdateFamilyTaskBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        completed:
          type: ["boolean", "null"]
        dueAt:
          type: ["string", "null"]
          format: date-time
        priority:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
    FamilyTask:
      type: object
      properties:
        id:
          type: integer
        familyId:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        completed:
          type: boolean
        dueAt:
          type: ["string", "null"]
          format: date-time
        priority:
          type: string
        assignedTo:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, familyId, userId, title, description, completed, dueAt, priority, assignedTo, createdAt]

    # ─── Progress ────────────────────────────────────────────────────────────
    CreateProgressBody:
      type: object
      properties:
        childId:
          type: integer
        module:
          type: string
          enum: [junior, student]
        subject:
          type: string
        score:
          type: integer
          minimum: 0
          maximum: 100
        notes:
          type: ["string", "null"]
      required: [childId, module, subject, score]
    Progress:
      type: object
      properties:
        id:
          type: integer
        childId:
          type: integer
        module:
          type: string
        subject:
          type: string
        score:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, childId, module, subject, score, notes, createdAt]