# lib/api-zod/src/generated/types/familyTask.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AYA Assistant API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FamilyTask {
  id: number;
  familyId: number;
  userId: number;
  title: string;
  /** @nullable */
  description: string | null;
  completed: boolean;
  /** @nullable */
  dueAt: Date | null;
  priority: string;
  /** @nullable */
  assignedTo: number | null;
  createdAt: Date;
}