# lib/api-zod/src/generated/types/updateFamilyTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AYA Assistant API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateFamilyTaskBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  completed?: boolean | null;
  /** @nullable */
  dueAt?: Date | null;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  assignedTo?: number | null;
}